+++ FicheroNacionalPip.Common/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace FicheroNacionalPip.Common;
-
-public class Class1
-{
-
-}