+++ FicheroNacionalPip.Presentation/Infrastructure/BooleanToVisibilityConverter.cs
@@ -1,0 +1,26 @@
+using System.Windows.Data;
+using System.Windows;
+using System.Globalization;
+
+namespace FicheroNacionalPip.Presentation.Infrastructure;
+
+public class BooleanToVisibilityConverter : IValueConverter
+{
+    public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+    {
+        if (value is bool boolValue)
+        {
+            return boolValue ? Visibility.Visible : Visibility.Collapsed;
+        }
+        return Visibility.Collapsed;
+    }
+
+    public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+    {
+        if (value is Visibility visibility)
+        {
+            return visibility == Visibility.Visible;
+        }
+        return false;
+    }
+}