+++ FicheroNacionalPip.Presentation/Models/MenuItemModel.cs
@@ -1,4 +1,6 @@
 using System.Collections.ObjectModel;
+using System.ComponentModel;
+using System.Runtime.CompilerServices;
 using System.Windows.Input;
 using MaterialDesignThemes.Wpf;
 
@@ -7,8 +9,10 @@ namespace FicheroNacionalPip.Presentation.Models;
 /// <summary>
 /// Representa un ítem de menú en la interfaz de usuario.
 /// </summary>
-public record class MenuItemModel
+public class MenuItemModel : INotifyPropertyChanged
 {
+    private bool _isEnabled = true;
+    
     /// <summary>
     /// Nombre a mostrar en el menú.
     /// </summary>
@@ -28,5 +32,28 @@ public record class MenuItemModel
     /// Colección de subítems del menú.
     /// </summary>
     public ObservableCollection<MenuItemModel>? SubItems { get; init; } = [];
+
+    /// <summary>
+    /// Indica si el ítem de menú está habilitado o no.
+    /// </summary>
+    public bool IsEnabled
+    {
+        get => _isEnabled;
+        set
+        {
+            if (_isEnabled != value)
+            {
+                _isEnabled = value;
+                OnPropertyChanged();
+            }
+        }
+    }
+    
+    public event PropertyChangedEventHandler? PropertyChanged;
+    
+    protected virtual void OnPropertyChanged([CallerMemberName] string? propertyName = null)
+    {
+        PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+    }
 }